+++ client-todo/src/todo/components/item.test.tsx
@@ -1,4 +1,3 @@
-
 import { expect, test, vi } from 'vitest'
 import { fireEvent, render } from '@testing-library/react'
 import Item, { Todo } from './item'
@@ -7,13 +6,14 @@ import '@testing-library/jest-dom';
 test('item component deletes item', () => {
 
   // create todo
-  const todo:Todo = { id: 1, title: 'Test Todo' };
+  const todo:Todo = { id: 1, title: 'Test Todo', description: 'Test Description', createdAt: '2021-01-01', updatedAt: '2021-01-01'};
 
   // mock delete function
   const mockDelete = vi.fn();
+  const mockSetCurrent = vi.fn();
 
   // render the component
-  const { getByTestId, getByRole } = render(<Item todo={todo} onDelete={mockDelete} />)
+  const { getByTestId, getByRole } = render(<Item setCurrentTodo={mockSetCurrent} todo={todo} onDelete={mockDelete} />)
 
   // test that the item is rendered
   expect(getByTestId('item-id').textContent).toBe(todo.id.toString());