+++ client-todo/src/todo/useTodo.ts
@@ -5,6 +5,7 @@ export const API_BASE_URL = ENV_URL;
 
 export const API_GET_ALL_TODOS = `${API_BASE_URL}/todos`;
 export const API_ADD_TODO = `${API_BASE_URL}/todo`;
+export const API_UPDATE_TODO = `${API_BASE_URL}/todo`;
 export const API_DELETE_TODO = `${API_BASE_URL}/todo`;
 
 export const useTodos = () => {
@@ -20,10 +21,15 @@ export const useTodos = () => {
         }
         return data;
     }
-    const { data, error, mutate, isLoading, isValidating } = useSWR(API_GET_ALL_TODOS, fetcher)
+    const getAllTodosOptions = {
+        //revalidateOnFocus: false, // stop page flicker 
+        //revalidateOnReconnect: false,
+        //revalidateOnMount: false,
+    };
+    const { data, error, mutate, isLoading, isValidating } = useSWR(API_GET_ALL_TODOS, fetcher,getAllTodosOptions)
 
 
-    const addTodo = async (newTodo: NewTodo): Promise<void> => {
+    const addTodo = async (newTodo: NewTodo) => {
         if (!newTodo) {
             console.log('addTodo', 'newTodo is undefined');
         }
@@ -34,32 +40,31 @@ export const useTodos = () => {
             },
             body: JSON.stringify({ todo: newTodo }),
         });
+        const returnedAddedTodo = await result.json()
         if (!result.ok) {
-            throw new Error(`result: ${result.status} ${result.statusText}`);
+            return returnedAddedTodo;
         }
-
-
-        const { data: returnedTodo } = await result.json();
-        console.log('addTodo', returnedTodo);
-        mutate([...data, returnedTodo]);
-        return returnedTodo;
+        mutate([...data, returnedAddedTodo?.data]);
+        return returnedAddedTodo;
     };
 
 
     const updateTodo = async (updatedTodo: Todo) => {
-        if (!data) {
-            console.log('updateTodo', 'id is undefined');
+        if (!updatedTodo.id) {
             return false;
         }
-        const response = await fetcher(API_ADD_TODO, { body: JSON.stringify({ todo: updatedTodo }) });
-        if (!response.ok) {
-            throw new Error('An error occurred while adding the todo.');
-        }
+        const options = { 
+            method: "PUT", 
+            headers: { "Content-Type": "application/json"},
+            body: JSON.stringify({ todo: updatedTodo }) };
+        const response = await fetch(`${API_UPDATE_TODO}/${updatedTodo.id}`, options);
         const returnedUpdatedTodo = await response.json()
-        console.log('updateTodo', returnedUpdatedTodo);
+        if (!returnedUpdatedTodo.ok) {
+            return returnedUpdatedTodo;
+        }     
         mutate(
             data.map(
-                (thisTodo: Todo) => (thisTodo.id === updatedTodo.id ? updatedTodo : thisTodo),
+                (thisTodo: Todo) => (thisTodo.id === returnedUpdatedTodo.data?.id ? returnedUpdatedTodo.data : thisTodo),
                 false
             )
         );
@@ -84,4 +89,4 @@ export const useTodos = () => {
     };
 
     return { data, error, addTodo, isLoading, updateTodo, removeTodo, isValidating };
-};
+}